Rename and set regex for name
